third_party/blink/web_tests/external/wpt/raw-sockets/remotePort-required.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Sockets test: open without remotePort</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      'use strict';

      const options = {
        localAddress: "127.0.0.1",
        localPort: 0,
        remoteAddress: "127.0.0.1",
      };

      test(() => {
        assert_throws_js(TypeError, () => { navigator.openTCPSocket() });
        assert_throws_js(TypeError, () => { navigator.openTCPSocket(null) });
        assert_throws_js(TypeError, () => { navigator.openTCPSocket(undefined) });
        assert_throws_js(TypeError, () => { navigator.openTCPSocket(options) });
      }, 'openTCPSocket requires remotePort');

      test(() => {
        assert_throws_js(TypeError, () => { navigator.openUDPSocket() });
        assert_throws_js(TypeError, () => { navigator.openUDPSocket(null) });
        assert_throws_js(TypeError, () => { navigator.openUDPSocket(undefined) });
        assert_throws_js(TypeError, () => { navigator.openTCPSocket(options) });
      }, 'openUDPSocket requires remotePort');
    </script>
  </body>
</html>
